[formal/conn] Support connectivity test locally
This PR adds basic configs to run connecivity test locally.
This is the first step to support running connecivity test using dvsim.
The reason we are adding this local option is because we can temporarily
using this script to run with GUI.

Next PR will support running the connectivity test on dvsim.

Signed-off-by: Cindy Chen <chencindy@google.com>