[topgen] Impose an ordering on blocks in top_uvm_reg.sv.tpl
Tracing things back through topgen.py, the insertion order of
top.blocks depends on the order that the glob returns filenames in
search_ips (in lib.py). This isn't guaranteed to be stable across
machines and the whole thing feels a bit delicate either way.

This commit uses the blocks in alphabetical order. Another option
would be to order by the base address of the first instance: maybe a
bit nicer, but more work so I've gone with the easy solution first.

To check this works on a single machine, add the line

    ips.reverse()

just after the definition of ips in search_ips (in lib.py). The
generate results before and after the addition, diffing the two

    mkdir -p X
    util/topgen.py -t hw/top_earlgrey/data/top_earlgrey.hjson -r -o X

With this patch, nothing changes.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>